requirements.txt
flask==1.1.1
python-dotenv~=0.10.1
flask-restplus==0.12.1
pytest-flask==0.10.0
sqlalchemy>=1.2.0b3
flask-sqlalchemy==2.3.2
Flask-Migrate==2.1.1
pymysql==0.9.3
pytest==3.6.2
py==1.5.3
pluggy==0.6.0
pytest-flask==0.10.0
pytest-cov==2.6.0
gunicorn==19.9.0
requests==2.18.4
Flask-Bcrypt==0.7.1
faker==1.0.4
sentry-sdk[flask]==0.7.7
Flask-Cors==3.0.8
PyJWT==1.4.0
pyOpenSSL==19.0.0

authlib==0.11
google-api-python-client
google-auth
virtualenv